fix: barrel import move exported directives in a specific file or in it's module
