+++ packages/context/src/lib/index.ts
@@ -1,20 +1,7 @@
-import { ContextImportExportComponent } from './context-import-export';
-import { CONTEXT_MANAGER_DIRECTIVES } from './context-manager';
-import { CONTEXT_MAP_BUTTON_DIRECTIVES } from './context-map-button';
-import { SHARE_MAP_DIRECTIVES } from './share-map';
-import { SidenavComponent } from './sidenav';
-
 export * from './context-import-export';
 export * from './context-manager';
 export * from './context-map-button';
 export * from './share-map';
 export * from './sidenav';
 export * from './environment';
-
-export const CONTEXT_DIRECTIVES = [
-  ContextImportExportComponent,
-  SidenavComponent,
-  ...CONTEXT_MANAGER_DIRECTIVES,
-  ...CONTEXT_MAP_BUTTON_DIRECTIVES,
-  ...SHARE_MAP_DIRECTIVES
-] as const;
+export * from './context.directive';
