+++ packages/geo/src/lib/catalog/catalog-library/catalog-library.module.ts
@@ -3,6 +3,11 @@ import { NgModule } from '@angular/core';
 import { AddCatalogDialogComponent } from './add-catalog-dialog.component';
 import { CatalogLibaryComponent } from './catalog-library.component';
 
+export const CATALOG_LIBRARY_DIRECTIVES = [
+  CatalogLibaryComponent,
+  AddCatalogDialogComponent
+] as const;
+
 /**
  * @deprecated import the CatalogLibaryComponent, AddCatalogDialogComponent directly
  */
