+++ packages/geo/src/lib/catalog/catalog-library/index.ts
@@ -1,10 +1,3 @@
-import { AddCatalogDialogComponent } from './add-catalog-dialog.component';
-import { CatalogLibaryComponent } from './catalog-library.component';
-
 export * from './add-catalog-dialog.component';
 export * from './catalog-library.component';
-
-export const CATALOG_LIBRARY_DIRECTIVES = [
-  CatalogLibaryComponent,
-  AddCatalogDialogComponent
-] as const;
+export * from './catalog-library.module';
