+++ packages/geo/src/lib/catalog/catalog.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 
-import { IgoCatalogBrowserModule } from './catalog-browser/catalog-browser.module';
-import { IgoCatalogLibraryModule } from './catalog-library/catalog-library.module';
+import { CatalogBrowserComponent } from './catalog-browser/catalog-browser.component';
+import { CATALOG_LIBRARY_DIRECTIVES } from './catalog-library';
+
+export const CATALOG_DIRECTIVES = [
+  CatalogBrowserComponent,
+  ...CATALOG_LIBRARY_DIRECTIVES
+] as const;
 
 /**
  * @deprecated import the components directly or the CATALOG_DIRECTIVES
  */
 @NgModule({
-  exports: [IgoCatalogBrowserModule, IgoCatalogLibraryModule]
+  imports: [...CATALOG_DIRECTIVES],
+  exports: [...CATALOG_DIRECTIVES]
 })
 export class IgoCatalogModule {}
