+++ packages/geo/src/lib/feature/feature.module.ts
@@ -1,12 +1,18 @@
 import { NgModule } from '@angular/core';
 
-import { IgoFeatureDetailsModule } from './feature-details/feature-details.module';
-import { IgoFeatureFormModule } from './feature-form/feature-form.module';
+import { FEATURE_DETAILS_DIRECTIVES } from './feature-details';
+import { FeatureFormComponent } from './feature-form/feature-form.component';
+
+export const FEATURE_DIRECTIVES = [
+  ...FEATURE_DETAILS_DIRECTIVES,
+  FeatureFormComponent
+] as const;
 
 /**
  * @deprecated import the components directly or the FEATURE_DIRECTIVES for the set
  */
 @NgModule({
-  exports: [IgoFeatureDetailsModule, IgoFeatureFormModule]
+  imports: [...FEATURE_DIRECTIVES],
+  exports: [...FEATURE_DIRECTIVES]
 })
 export class IgoFeatureModule {}
