+++ packages/geo/src/lib/feature/index.ts
@@ -1,11 +1,4 @@
-import { FEATURE_DETAILS_DIRECTIVES } from './feature-details';
-import { FeatureFormComponent } from './feature-form/feature-form.component';
-
 export * from './shared';
+export * from './feature.module';
 export * from './feature-details';
 export * from './feature-form/feature-form.component';
-
-export const FEATURE_DIRECTIVES = [
-  ...FEATURE_DETAILS_DIRECTIVES,
-  FeatureFormComponent
-] as const;
