+++ packages/geo/src/lib/filter/filter.module.ts
@@ -24,7 +24,7 @@ import { TimeFilterItemComponent } from './time-filter-item/time-filter-item.com
 import { TimeFilterListBindingDirective } from './time-filter-list/time-filter-list-binding.directive';
 import { TimeFilterListComponent } from './time-filter-list/time-filter-list.component';
 
-const DIRECTIVES = [
+export const FILTER_DIRECTIVES = [
   FilterableDataSourcePipe,
   TimeFilterButtonComponent,
   TimeFilterFormComponent,
@@ -43,13 +43,14 @@ const DIRECTIVES = [
   SpatialFilterItemComponent,
   OgcFilterTimeComponent,
   OgcFilterTimeSliderComponent
-];
+] as const;
+
 /**
  * @deprecated import the components directly or the FILTER_DIRECTIVES for the set
  */
 @NgModule({
-  imports: DIRECTIVES,
-  exports: DIRECTIVES,
+  imports: [...FILTER_DIRECTIVES],
+  exports: [...FILTER_DIRECTIVES],
   providers: [
     TimeFilterService,
     OGCFilterService,
