+++ packages/geo/src/lib/geometry/geometry-form-field/geometry-form-field.module.ts
@@ -3,6 +3,11 @@ import { NgModule } from '@angular/core';
 import { GeometryFormFieldInputComponent } from './geometry-form-field-input.component';
 import { GeometryFormFieldComponent } from './geometry-form-field.component';
 
+export const GEOMETRY_FORM_FIELD_DIRECTIVES = [
+  GeometryFormFieldComponent,
+  GeometryFormFieldInputComponent
+] as const;
+
 /**
  * @deprecated import the components directly or the GEOMETRY_FORM_FIELD_DIRECTIVES for the set
  */
