+++ packages/geo/src/lib/geometry/geometry-form-field/index.ts
@@ -1,10 +1,3 @@
-import { GeometryFormFieldInputComponent } from './geometry-form-field-input.component';
-import { GeometryFormFieldComponent } from './geometry-form-field.component';
-
-export * from './geometry-form-field-input.component';
+export * from './geometry-form-field.module';
 export * from './geometry-form-field.component';
-
-export const GEOMETRY_FORM_FIELD_DIRECTIVES = [
-  GeometryFormFieldComponent,
-  GeometryFormFieldInputComponent
-] as const;
+export * from './geometry-form-field-input.component';
