+++ packages/geo/src/lib/import-export/index.ts
@@ -1,13 +1,4 @@
-import { ExportButtonComponent } from './export-button';
-import { ImportExportComponent } from './import-export';
-import { DropGeoFileDirective } from './shared';
-
 export * from './shared';
+export * from './import-export.module';
 export * from './import-export';
 export * from './export-button';
-
-export const IMPORT_EXPORT_DIRECTIVES = [
-  ImportExportComponent,
-  DropGeoFileDirective,
-  ExportButtonComponent
-] as const;
