+++ packages/geo/src/lib/metadata/index.ts
@@ -1,12 +1,3 @@
-import {
-  MetadataAbstractComponent,
-  MetadataButtonComponent
-} from './metadata-button';
-
 export * from './shared';
+export * from './metadata.module';
 export * from './metadata-button';
-
-export const METADATA_DIRECTIVES = [
-  MetadataButtonComponent,
-  MetadataAbstractComponent
-] as const;
