+++ packages/geo/src/lib/search/search-results/search-results.directive.ts
@@ -1,0 +1,7 @@
+import { SearchResultAddButtonComponent } from './search-results-add-button.component';
+import { SearchResultsComponent } from './search-results.component';
+
+export const SEARCH_RESULTS_DIRECTIVES = [
+  SearchResultsComponent,
+  SearchResultAddButtonComponent
+] as const;
