+++ packages/integration/src/lib/context/index.ts
@@ -1,17 +1,6 @@
-import { ContextEditorToolComponent } from './context-editor-tool/context-editor-tool.component';
-import { ContextManagerToolComponent } from './context-manager-tool/context-manager-tool.component';
-import { ContextPermissionManagerToolComponent } from './context-permission-manager-tool/context-permission-manager-tool.component';
-import { ContextShareToolComponent } from './context-share-tool/context-share-tool.component';
-
 export * from './context-manager-tool/context-manager-tool.component';
 export * from './context-editor-tool/context-editor-tool.component';
 export * from './context-permission-manager-tool/context-permission-manager-tool.component';
 export * from './context-share-tool/context-share-tool.component';
 export * from './context.state';
-
-export const INTEGRATION_CONTEXT_DIRECTIVES = [
-  ContextEditorToolComponent,
-  ContextManagerToolComponent,
-  ContextPermissionManagerToolComponent,
-  ContextShareToolComponent
-] as const;
+export * from './context.directive';
