+++ packages/integration/src/lib/map/index.ts
@@ -1,10 +1,3 @@
-import { AdvancedMapToolComponent } from './advanced-map-tool';
-import { MapDetailsToolComponent } from './map-details-tool';
-import { MapLegendToolComponent } from './map-legend';
-import { MapProximityToolComponent } from './map-proximity-tool';
-import { MapToolComponent } from './map-tool';
-import { MapToolsComponent } from './map-tools';
-
 export * from './advanced-map-tool';
 export * from './map-proximity-tool';
 export * from './map-details-tool';
@@ -13,12 +6,4 @@ export * from './map-tool';
 export * from './map-tools';
 export * from './map.state';
 export * from './map-proximity.state';
-
-export const INTEGRATION_MAP_DIRECTIVES = [
-  AdvancedMapToolComponent,
-  MapProximityToolComponent,
-  MapToolComponent,
-  MapToolsComponent,
-  MapDetailsToolComponent,
-  MapLegendToolComponent
-] as const;
+export * from './map.directive';
