+++ packages/integration/src/lib/search/index.ts
@@ -1,14 +1,6 @@
-import { SearchBarBindingDirective } from './search-bar/search-bar-binding.directive';
-import { SearchResultsToolComponent } from './search-results-tool/search-results-tool.component';
-
 export * from './query.state';
 export * from './search.state';
 export * from './search-results-tool/search-results-tool.module';
 export * from './search-results-tool/search-results-tool.component';
 export * from './search-bar/search-bar.module';
 export * from './search-bar/search-bar-binding.directive';
-
-export const INTEGRATION_SEARCH_DIRECTIVES = [
-  SearchBarBindingDirective,
-  SearchResultsToolComponent
-] as const;
